fix case-error, use static_casts

+++ tinyexr.h
@@ -651,7 +651,7 @@ extern int LoadEXRFromMemory(float **out_rgba, int *width, int *height,
 #ifndef NOMINMAX
 #define NOMINMAX
 #endif
-#include <windows.h>  // for UTF-8 and memory-mapping
+#include <Windows.h>  // for UTF-8 and memory-mapping
 
 #if !defined(WINAPI_FAMILY) || (WINAPI_FAMILY == WINAPI_FAMILY_DESKTOP_APP)
 #define TINYEXR_USE_WIN32_MMAP (1)
@@ -4421,11 +4421,11 @@ static unsigned char **AllocateImage(int num_channels,
 
 #ifdef _WIN32
 static inline std::wstring UTF8ToWchar(const std::string &str) {
-  int wstr_size =
-      MultiByteToWideChar(CP_UTF8, 0, str.data(), (int)str.size(), NULL, 0);
-  std::wstring wstr(wstr_size, 0);
-  MultiByteToWideChar(CP_UTF8, 0, str.data(), (int)str.size(), &wstr[0],
-                      (int)wstr.size());
+  int wstr_size = MultiByteToWideChar(CP_UTF8, 0, str.data(),
+                                      static_cast<int>(str.size()), NULL, 0);
+  std::wstring wstr(static_cast<size_t>(wstr_size), 0);
+  MultiByteToWideChar(CP_UTF8, 0, str.data(), static_cast<int>(str.size()),
+                      &wstr[0], static_cast<int>(wstr.size()));
   return wstr;
 }
 #endif
